Add basic frame timing to main loop
I've tried to be conscious of a few issues here -- timer overflow,
frames that are too short, and so on.  The use of floats for elapsed
time might be hazardous, precision-wise.

A static local variable is used to simplify keeping track of ticks
between frames.  This routine is only ever called from the main
scheduling loop, in a single thread.