Make vertex buffer attribute a singleton
Following sprite's lead, it seems to me that we don't need one of these
for every tilemap or font, since we're setting it up for every item.
Maybe that will change with more clever batching, but for now, get rid
of it.

The initialization changes in tilemap are due to gcc bug 53119:
  https://gcc.gnu.org/bugzilla/show_bug.cgi?id=53119